refactor: extract GitHub provider, mark as deprecated
- Move GitHub provider implementation to src/vcs/github/provider.ts
- Create GitHubProvider class implementing IVCSProvider
- Keep existing CLI commands in src/vcs/github/github.ts
- Re-export from src/cli/cmd/github.ts for backward compatibility
- Add @deprecated JSDoc comments
- Create tests for GitHub provider
